Clear metrics from starlette_exporter as well
